Interpreter: Fix and document substitute
With b000c2900ccfb2bedef4abac6fd99f56d5b9007a we now need to be careful
how deep we traverse the structure when substituting a variable.

This function is a bit of a mess, and we really should be using De
Bruijn indices, but that's a TODO for a later day.