Differentiate between checked and unchecked exprs at the type level
This is a very nice use case for phantom types!  This also allows us to
partially specify the "standard library" functions (like +, -, etc.)
without running the risk of them breaking at runtime, as well as get rid
of a (now superfluous) exception.

I do wonder whether at this point one should not move away from
exceptions to something like Either MyErrors a, but for this small of a
language the ergonomics seem to outweigh the costs (at least for now).